2005-02-17	Ralf Corsepius <ralf.corsepius@rtems.org>
	* libblock/include/rtems/ramdisk.h: size_t
	rtems_ramdisk_configuration_size.
	* libblock/src/ramdisk.c: Adaptations for 16bit target compliance.

+++ cpukit/libblock/include/rtems/ramdisk.h
@@ -35,7 +35,7 @@ typedef struct rtems_ramdisk_config {
  * The following is definitions for RAM disk configuration table
  */
 extern rtems_ramdisk_config rtems_ramdisk_configuration[];
-extern int rtems_ramdisk_configuration_size;
+extern size_t rtems_ramdisk_configuration_size;
 
 /* ramdisk_initialize --
  *     RAM disk driver initialization entry point.
